move files to start reorganizing by feature
