+++ src/app/services/tutor/utils.py
@@ -5,7 +5,7 @@ from fastapi import HTTPException, UploadFile
 from pypdf import PdfReader
 from qdrant_client.models import ScoredPoint
 
-from src.app.api.dependencies import get_settings
+from src.app.shared.utils.dependencies import get_settings
 from src.app.services.pdf_extractor import extract_txt_from_pdf_with_tika
 from src.app.utils.decorators import log_time_and_error_sync
