+++ src/main.py
@@ -12,7 +12,7 @@ from qdrant_client.http import exceptions as qdrant_exceptions
 from starlette.exceptions import HTTPException as StarletteHTTPException
 
 from src.app.api.api_v1.api import api_router, api_tags_metadata
-from src.app.api.shared.enpoints import health
+from src.app.shared.api import health
 from src.app.core.config import settings
 from src.app.core.lifespan import lifespan
 from src.app.middleware.monitor_requests import MonitorRequestsMiddleware
